Matched the scala key generation, command format.
The scala version was recently updated to generate a stronger
key (more iterations of the hash to make it take longer, and
also cover more bits.  Also, I like that the scala version is
only one commmand with hash and crypt subcommands.  So, I altered
the Go version to match it.